Moved DTOs from single file to separate files per DTO
