+++ backend/FlashIdBackend/Application/Features/Auth/DTOs/LoginRequestDto.cs
@@ -1,0 +1,7 @@
+namespace Application.Features.Auth.DTOs;
+
+public class LoginRequestDto
+{
+    public string Email { get; set; } = string.Empty;
+    public string Password { get; set; } = string.Empty;
+}
