+++ backend/FlashIdBackend/Application/Features/Auth/DTOs/LoginResponseDto.cs
@@ -1,11 +1,5 @@
 namespace Application.Features.Auth.DTOs;
 
-public class LoginRequestDto
-{
-    public string Email { get; set; } = string.Empty;
-    public string Password { get; set; } = string.Empty;
-}
-
 public class LoginResponseDto
 {
     public string Token { get; set; } = string.Empty;
@@ -14,9 +8,4 @@ public class LoginResponseDto
     public string Role { get; set; } = string.Empty;
     public string Names { get; set; } = string.Empty;
     public string Surname { get; set; } = string.Empty;
-}
-
-public class LogoutResponseDto
-{
-    public string Message { get; set; } = string.Empty;
 }
