+++ backend/FlashIdBackend/Application/Features/Auth/DTOs/LogoutResponseDto.cs
@@ -1,0 +1,6 @@
+namespace Application.Features.Auth.DTOs;
+
+public class LogoutResponseDto
+{
+    public string Message { get; set; } = string.Empty;
+}
